pyproject.toml
[project]
name = "scene-understanding-ai-judge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.5.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "rich>=14.2.0",
    "streamlit>=1.50.0",
]
